infodeck: migrate jasmine tests to mocha
some tests disabled while I investigate issues

trigger for this was running jasmine tests and find that some fail now
(some non-deterministically) that were ok when last run a year or so
ago. Hypothesis is that there is some bug in the async handling as I
don't see problem when manually working with deck. Certainly there is
some issues due to jquery's handling of promises (eg bug fixed in
7826f15)